refactor: move new services in dedicated root
