+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/storage/application/adapter/ApplicationMapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.storage.application.adapter;
+package io.gravitee.apim.storage.application.adapter;
 
 import io.gravitee.definition.model.Origin;
 import io.gravitee.repository.management.model.Application;
